Use float32 again instead of float64. Precision error caused by float32 previously was due to using float32 epsilon in _safe_divide.

+++ src/psd_export/composite.py
@@ -14,7 +14,7 @@ from . import blendfuncs
 from . import util
 from .util import peval
 
-dtype = np.float64
+dtype = np.float32
 
 class WrappedLayer():
     def __init__(self, layer:Layer, clip_layers=[], parent:WrappedLayer=None):
